VPN NetBird -> Delete NetBird-Peer rather than NetBird-Group when generating Setup-String anew.
This means that re-generating a Setup-String does not fail anymore,
It used to fail, because it is not possible to delete a NetBird-Group
that's associated with a NetBird-Setup-Key (and it doesn't seem possible to delete a NetBird-Setup-Key).

This also means that the NetBird Peer gets logged out,
which allows the next EDGAR Setup to log it back in,
with which it can adjust the MTU, if necessary.